LayoutTests/http/tests/security/contentSecurityPolicy/combine-multiple-policies-expected.txt

CONSOLE MESSAGE: line 11: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self'".

CONSOLE MESSAGE: line 14: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'none'".

This test checks that we enforce all the supplied policies. This test passes if it doesn't alert fail and if the style doesn't apply.
